Revert "feat: 좌표를 Int 형으로 변경"
This reverts commit 47b75efddbe7ef2b4b79383fb67b3ae9c78a5de2.

+++ app/src/main/java/com/example/timemacro/ClickerAccessibilityService.kt
@@ -12,11 +12,9 @@ class ClickerAccessibilityService : AccessibilityService() {
 
     override fun onInterrupt() {}
 
-    // performClick 함수 수정: Float 형 x, y를 Int 형으로 받도록 변경
-    fun performClick(x: Int, y: Int): Boolean {
-        // Int 형 x, y를 Float 형으로 변환하여 Path 객체 생성에 사용
+    fun performClick(x: Float, y: Float): Boolean {
         val path = Path()
-        path.moveTo(x.toFloat(), y.toFloat())
+        path.moveTo(x, y)
 
         val gesture = GestureDescription.Builder()
             .addStroke(GestureDescription.StrokeDescription(path, 0, 100))
